+++ R/f_summarise.R
@@ -88,6 +88,7 @@ f_summarise <- function(data, ..., .by = NULL,
 
   group_vars <- get_groups(temp_data, {{ .by }})
 
+  # Groups for collapse functions
   no_groups <- length(group_vars) == 0
   if (no_groups){
     groups <- NULL
@@ -99,9 +100,7 @@ f_summarise <- function(data, ..., .by = NULL,
     )
   }
 
-
-  ## Flags so that we only construct grouped_df if we need to and do it once
-  # construct_grouped_df <- FALSE
+  ## Flag so that we only construct grouped_df if we need to and do it once
   grouped_df_has_been_constructed <- FALSE
 
   dots <- rlang::enquos(...)
@@ -125,11 +124,13 @@ f_summarise <- function(data, ..., .by = NULL,
   for (i in seq_along(dots)){
     dot <- dots[[i]]
     dot_label <- rlang::as_label(dot)
+    dot_env <- rlang::quo_get_env(dot)
+    dot_expr <- rlang::quo_get_expr(dot)
     dot_nm <- dot_nms[i]
+
     if (!nzchar(dot_nm)){
       dot_nm <- dot_label
     }
-    dot_env <- rlang::quo_get_env(dot)
     if (rlang::quo_is_call(dot)){
       dot_args <- rlang::call_args(dot)
       if (length(dot_args) == 0){
@@ -137,8 +138,8 @@ f_summarise <- function(data, ..., .by = NULL,
       } else {
         var <- as.character(dot_args[[1]])
       }
-    }  else {
-      var <- rlang::as_label(rlang::quo_get_expr(dot))
+    } else {
+      var <- rlang::as_label(dot_expr)
     }
     var_not_nested <- length(var) <= 1
     if (.optimise && is_n_call(dot) && var_not_nested){
@@ -159,7 +160,8 @@ f_summarise <- function(data, ..., .by = NULL,
           fun, c(list(temp_data[[var]],
                       g = groups,
                       use.g.names = FALSE),
-                 fun_args)
+                 fun_args),
+          envir = dot_env
         )
         out[[dot_nm]] <- res
 
@@ -175,9 +177,9 @@ f_summarise <- function(data, ..., .by = NULL,
       dot_args <- rlang::call_args(dot)
       across_expr <- match.call(
         definition = dplyr::across,
-        call = rlang::quo_get_expr(dot),
+        call = dot_expr,
         expand.dots = FALSE,
-        envir = rlang::quo_get_env(dot)
+        envir = dot_env
       )
 
       if (!".cols" %in% names(across_expr)){
@@ -220,32 +222,42 @@ f_summarise <- function(data, ..., .by = NULL,
         fns <- rlang::as_label(across_fns)
         fn_names <- fns
       }
-      fn_matches <- cheapr::na_rm(sort(c(match(fns, base_fns),
-                                         match(fns, collapse_fns))))
-      which_fns <- which(fns %in% base_fns | fns %in% collapse_fns)
-      which_other_fns <- which(fns %in% base_fns | fns %in% collapse_fns,
-                               invert = TRUE)
-      fast_fn_names <- collapse_fns[fn_matches]
+      base_matches <- match(fns, base_fns)
+      collapse_matches <- match(fns, collapse_fns)
+      fast_fns <- rep_len(NA_character_, length(fns))
+      for (i in seq_along(fns)){
+        if (!is.na(base_matches[i])){
+          fast_fns[i] <- collapse_fns[base_matches[i]]
+        } else if (!is.na(collapse_matches[i])){
+          fast_fns[i] <- collapse_fns[collapse_matches[i]]
+        }
+      }
+      which_fast <- cheapr::na_find(fast_fns, invert = TRUE)
+      which_other <- cheapr::na_find(fast_fns)
+      fast_fns <- cheapr::na_rm(fast_fns)
+
       full_res <- vector("list", length(vars) * length(fns))
       col_matrix <- matrix(logical( length(vars) * length(fns)),
-                           nrow = length(vars),
-                           ncol = length(fns))
-      col_matrix[, which_fns] <- TRUE
-      across_res <- fast_eval_across(temp_data, groups, vars, fast_fn_names, dot_env)
+                           nrow = length(fns),
+                           ncol = length(vars))
+      for (col in seq_along(vars)){
+        col_matrix[which_fast, col] <- TRUE
+      }
+      across_res <- fast_eval_across(temp_data, groups, vars, fast_fns, dot_env)
 
-      if (length(which_other_fns) > 0){
+      if (length(which_other) > 0){
+        if (!grouped_df_has_been_constructed){
+          temp_data <- construct_grouped_df(temp_data, groups, group_vars)
+        }
+        grouped_df_has_been_constructed <- TRUE
         if (across_fns_as_list){
           dplyr_res <- dplyr::summarise(
             temp_data, dplyr::across(
               dplyr::all_of(vars),
-              rlang::eval_tidy(across_fns, env = dot_env)[which_other_fns]
+              rlang::eval_tidy(across_fns, env = dot_env)[which_other]
             ), .groups = "drop"
           )
         } else {
-          if (!grouped_df_has_been_constructed){
-            temp_data <- construct_grouped_df(temp_data, groups, group_vars)
-          }
-          grouped_df_has_been_constructed <- TRUE
           dplyr_res <- dplyr::summarise(
             temp_data, dplyr::across(
               dplyr::all_of(vars),
@@ -291,12 +303,17 @@ fast_eval_across <- function(data, g, .cols, .fns, env, .names = NULL){
   ncols <- length(.cols)
   nfns <- length(.fns)
   out <- vector("list", ncols * nfns)
+  collapse_ns <- asNamespace("collapse")
   i <- 1L
   for (col in .cols){
     for (f in .fns){
-     fun <- get_from_package(f, "collapse")
+     fun <- get(f, collapse_ns, inherits = FALSE)
      var <- .subset2(data, col)
-     out[[i]] <- do.call(fun, list(var, g = g, use.g.names = FALSE), envir = env)
+     res <- do.call(fun, list(var, g = g, use.g.names = FALSE), envir = env)
+     if (length(var) == 0 && (identical(fun, collapse::fsd) || identical(fun, collapse::fvar))){
+       res <- numeric()
+     }
+     out[[i]] <- res
      i <- i + 1L
    }
   }
@@ -304,26 +321,30 @@ fast_eval_across <- function(data, g, .cols, .fns, env, .names = NULL){
 }
 
 across_col_names <- function (.cols = NULL, .fns = NULL, .names = NULL){
+  fns_null <- is.null(.fns)
   nms_null <- is.null(.names)
-  if (nms_null && length(.fns) == 1L) {
+  n_fns <- length(.fns)
+  n_cols <- length(.cols)
+
+  if (fns_null && nms_null){
+    out <- as.character(.cols)
+  } else if (nms_null && n_fns == 1L) {
     out <- .cols
-  } else if (nms_null && length(.cols) == 1L) {
+  } else if (nms_null && n_cols == 1L) {
     out <- .fns
   } else {
-    out <- character(length(.cols) * length(.fns))
+    out <- character(n_cols * n_fns)
     init <- 0L
-    .fn <- .fns
     if (nms_null) {
       for (.col in .cols) {
-        out[seq_along(.fns) + init] <- stringr::str_c(.col,
-                                                      "_", .fn)
-        init <- init + length(.fns)
+        out[seq_len(n_fns) + init] <- paste0(.col, "_", .fns)
+        init <- init + n_fns
       }
-    }
-    else {
+    } else {
+      .fn <- .fns
       for (.col in .cols) {
-        out[seq_along(.fns) + init] <- stringr::str_glue(.names)
-        init <- init + length(.fns)
+        out[seq_len(n_fns) + init] <- stringr::str_glue(.names)
+        init <- init + n_fns
       }
     }
   }